requirements.txt
httpx>=0.27.0
cryptography>=42.0.0
PyYAML>=6.0.0
